chore: centralize project type definitions
