+++ src/app/types/forms.ts
@@ -1,0 +1,4 @@
+import type { z } from 'zod';
+import { formSchema } from '@/app/components/agendamento/schemas';
+
+export type ScheduleFormValues = z.infer<typeof formSchema>;
